Refactor ProjectPanel Vue Component
