+++ resources/js/components/Profile.vue
@@ -1,4 +1,4 @@
-<template>
+  <template>
 	<div>
 		   <div class="page-top">
 		   	 <span>
