+++ resources/js/components/ProjectFeature/ProjectFeature.vue
@@ -50,7 +50,7 @@
 export default {
 
   components: {ProjectUpdate,ProjectSms,ProjectMail},
-
+ 
     props:['project','subscribe','members'],
     data() {
         return {
@@ -124,18 +124,6 @@ then(response=>{
   console.log(error.response.data.errors);
 });
 },
-
-  sweetAlert($action){
-    return swal.fire({
-      title: 'Are you sure?',
-      text: "You can be able to revert this!",
-      type: 'warning',
-      showCancelButton: true,
-      confirmButtonColor: '#3085d6',
-      cancelButtonColor: '#d33',
-      confirmButtonText: $action
-    })
-   },
 
    redirectSuccess($redirect){
      swal.fire(
